requirements.txt
openai==1.101.0
python-dotenv==1.1.1
vk_api==11.9.9
